backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
yt-dlp==2023.11.16
youtube-transcript-api==0.6.1
requests==2.31.0
beautifulsoup4==4.12.2
httpx==0.25.2
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
google-api-python-client==2.108.0
duckduckgo-search==4.1.1
numpy==1.24.3
wikipedia==1.4.0

